app/requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
langchain>=0.1.12
langchain-community>=0.0.27
langchain-groq>=0.2.0
python-dotenv>=1.0.1
huggingface-hub>=0.21.4
requests>=2.31.0